Adds initial version of the groundwater policy model
I am translating the legacy code of the campaspe integrated model.
Because of its simplicity I decided to start with the groundwater policy
model. This code adds this translation, including a method to create the
gw_state (groundwater_state) struct and run one iteration of the
groundwater model. More tests still need to be added.