Replace current_time to ts in SwState and EnvironmentState
During the translation it was decided to change the variable name ts of
the surface water state to current_time, to improve code readability.
However, multiple other places in the code are still using ts. That
makes reading the code more difficult. To maintain a better consitency
in the code we replaced current_time to ts in the SwState and
EnvironmentState.