Fix shared reference on sw_state (policy) setup
The bug was caused by the add_info dictionary that was being used on
all zones. All zones had a shared add_info dictionary causing its values
to be shared. We fixed this by creating a new function that creates a
new dictionary for each zone.

This commit also fix a simpler bug in the sw_allocation.jl. It was
checking for ts == 0, but ts is one-index and is never zero.

Besides this two fix, we refactor the variable names on sw_state, and
made minor refactors on environmental release dates check and
surface_water_update of the policy model.