Temporary fix on high groundwater use
The current setup has different allocation and entitlements per zone for
the policy and farm model. This result in the farm model being capable
of setting orders higher than the total allocated water in the policy
model. As a temporary fix to avoid the @assert error, we set the
groundwater use to be the allocation.