Change dam_extraction setup and update
The dam_extraction dataframe is defined in the policy model, sw_state
and updated during the main time loop. When using past scenarios we use
a dam_extraction data frame, but for simulating the future, we create an
empty data frame. We also decouple the two dataframes by adding a
deepcopy.