Remove GeoDataFrames and BlackBoxOptim from deps
GeoDataFrames was being used to load shapefiles but as this was not
being used it was removed. This commit updates the code to remove the
package and the code loading a GeoDataFrame was replaced by standard
.csv and dataframe.

BlackBoxOptim was never used in the core package and will now be
used only inside the sandbox.