Fix farm model not running on season_end
In the current setup when running the code for a long period, only one
year was showed in the results. This was happening because the Agtor.jl
code only output results if the model is ran on the season_end.
We initially supposed the season_end is fixed for all crops
and years, but apparently it can change. This commit updates the
plan_season_end in the begginning of the season to guarantee that the
farm model will always run in the end of the season and compute profits.