Implements method to adopt option
At the moment it is only possible to pass one option and not a sequence
of options. The option is adopted just before the time loop. Later the
farm_option can be changed to receive a sequence of option, in this case
the current if will adopt the first option and during the main time loop
the option will be changed.