Update policy_state entitlement computation
The python code does different process of the original shapefiles to
compute entitlements per zone. Part is done in main.py and part in the
initialisation of the policy model. The julia code was including all
this process in the policy model setup. We moved part of the
entitlements computation pre-simulaiton and updated the julia policy
model setup to have the same behaviour as the python code.
(Data on entitlements passed to Agtor was also updated to follow the
same results of the results passed to the policy model).