Fix bug of gw_used > gw_alloc
In some contexts the volume of used groundwater was passing the volume
of allocated groundwater. The problem was being caused because the
used_water was computed incorrectly. The farm update was outputing
cumulative used water instead of volume used in the last fortnight. The
farm model was updated to return the correct value.