Update farm_model to apply rainfall correctly
The original famr model run fortnightly but the new farm model is
designed to run daily. A bug was happening because the irrigation was
never happening. This was caused because the rainfall and
evapotranspiration data was only being applied for one day of the
fortnight, not all the period. To solve this we manually apply the
rainfall and evapotranspiration for days out of the fortnight.

We also fix two small bugs that happened after fixing this. Sometimes
there was a negative groundwater allocation and the new season start was
not being updated correctly.

With this new changes the irrigation is happening.